% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibufunction.R
\name{kol_name}
\alias{kol_name}
\title{GENERALIZED KOL_NAME FUNCTION}
\usage{
kol_name(df)
}
\arguments{
\item{df}{- Dataframe to be formatted.}
}
\description{
\preformatted{
This is a function to format column names by :-
 1. lowering the case of the column names.
 2. trimming whitespace.
 3. substituting '.' with spaces in the columns.
 }
}
