% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibufunction.R
\name{numeric_checker}
\alias{numeric_checker}
\title{GENERALIZED NUMERIC_CHECKER FUNCTION}
\usage{
numeric_checker(df, kelas, negative = TRUE, na = TRUE, nan = TRUE, null = TRUE)
}
\arguments{
\item{df}{- Dataframe to be checked.}

\item{kelas}{- Classes of the columns.}

\item{negative}{- If TRUE, will run the is_non_negative function.}

\item{na}{- If TRUE, will run the is_not_na function.}

\item{nan}{- If TRUE, will run the is_not_nan function.}

\item{null}{- If TRUE, will run the is_not_null function.}
}
\value{
This function will return a dataframe with the numeric_checker_list list.
}
\description{
This is a function to find all NA, NAN, NULL, and negative value inside numeric classes of dataframe.
}
\author{
IBU Analytics
}
