% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibufunction.R
\name{path_checker}
\alias{path_checker}
\title{GENERALIZED PATH_CHECKER FUNCTION}
\usage{
path_checker(path)
}
\arguments{
\item{path}{= The path in which to be checked and created.}
}
\value{
This function will not return an output,

it will just create a folder in the directory path given if it does not exist.
}
\description{
This is a function to check and create path if doesn't exist
}
\author{
IBU Analytics
}
